# src/main/resources/openapi/catalog.yaml
openapi: 3.0.3
info:
  title: Catalog API
  version: 1.0.0
tags:
  - name: Products
    description: 商品目錄管理

paths:
  /products:
    get:
      tags: [Products]
      operationId: getProducts
      summary: 列出所有商品
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
    post:
      tags: [Products]
      operationId: createProduct
      summary: 新增商品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: 商品已建立
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"

  /products/{id}:
    get:
      tags: [Products]
      operationId: getProductById
      summary: 取得單一商品
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 商品資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: 商品不存在

components:
  schemas:
    CreateProductRequest:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          $ref: "#/components/schemas/Money"

    ProductResponse:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          $ref: "#/components/schemas/Money"

    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: decimal
        currency:
          type: string
          example: TWD